+++ src/BaseElement.ts
@@ -173,12 +173,22 @@ export class BaseElement extends HTMLElement {
     https://developer.mozilla.org/en-US/docs/Web/API/HTML_DOM_API/Microtask_guide
     */
     queueMicrotask(() => {
+      /* Apply all attributes */
+      this.applyAllAttributes()
+
       /* Invoke mount method */
       this.mountedCallback()
 
       /* Emit ready event */
       this.dispatchEvent(new CustomEvent("ready", { bubbles: true, cancelable: false }))
     })
+  }
+
+  applyAllAttributes() {
+    const attributes = this.getAllAttributes()
+    for (const key in attributes) {
+      this.attributeChangedCallback(key, "", attributes[key])
+    }
   }
 
   disconnectedCallback() {