+++ src/util/applyProps.ts
@@ -1,5 +1,8 @@
 import { IStringIndexable } from "../types"
 import { camelize } from "./camelize"
+import { getThreeObjectBySelector } from "./getThreeObjectBySelector"
+import { attributeValueToArray } from "./attributeValueToArray"
+import { parseDeg } from "./parseDeg"
 
 const IGNORED_KEYS = ["id"]
 
@@ -10,66 +13,87 @@ export const applyProps = (object: IStringIndexable, props: IStringIndexable) =>
 
     const key = camelize(firstKey)
 
-    /* Attempt to parse the value */
-    let parsed = undefined
-    try {
-      parsed = JSON.parse(value)
-    } catch (e) {}
+    /* Skip all ignored keys. */
+    if (IGNORED_KEYS.includes(key)) return
 
-    switch (true) {
-      /* Skip ignored keys */
-      case IGNORED_KEYS.includes(key):
-        break
+    /* Skip all data attributes. */
+    if (firstKey.startsWith("data-")) return
 
-      /* Ignore all data- keys */
-      case firstKey.startsWith("data-"):
-        break
-
-      /* Handle nested keys, ie. position-x */
-      case key in object && rest.length > 0:
-        applyProps(object[key], { [rest.join(".")]: value })
-        break
-
-      /*
-      Handle boolean properties. We will check against the only values that we consider falsey here,
-      taking into account that they might be coming from string-based HTML element attributes, where a
-      stand-alone boolean attribute like "cast-shadow" will emit a value of "". Eh!
-      */
-      case typeof object[key] === "boolean":
-        object[key] = ![undefined, null, false, "no", "false"].includes(value)
-        break
-
-      /* Handle properties that provide .set methods */
-      case object[key]?.set !== undefined:
-        switch (true) {
-          /* If the value is an array, feed its destructured representation to the set method. */
-          case Array.isArray(parsed):
-            object[key].set(...parsed)
-            break
-
-          /* A bit of special handling for "scale" properties, where we'll also accept a single numerical value */
-          case key === "scale" && typeof parsed === "number":
-            object[key].setScalar(parsed)
-            break
-
-          /* If we have a parsed value, set it directly */
-          case parsed:
-            object[key].set(parsed)
-            break
-
-          /* Otherwise, set the original string value, but split by commas */
-          default:
-            const list = value.split(",").map((el: string) => parseFloat(el) || el)
-            object[key].set(...list)
-        }
-        break
-
-      default:
-        /*
-        If we've reached this point, we're finally able to set a property on the object.
-        Amazing! But let's only do it if the property key is actually known.
-        */
-        if (key in object) object[key] = parsed !== undefined ? parsed : value
+    /* Recursively handle nested keys, eg. position.x */
+    if (key in object && rest.length > 0) {
+      applyProps(object[key], { [rest.join(".")]: value })
+      return
     }
+
+    /*
+    Handle boolean properties. We will check against the only values that we consider falsey here,
+    taking into account that they might be coming from string-based HTML element attributes, where a
+    stand-alone boolean attribute like "cast-shadow" will emit a value of "". Eh!
+    */
+    if (typeof object[key] === "boolean") {
+      object[key] = ![undefined, null, false, "no", "false"].includes(value)
+      return
+    }
+
+    /* It is attribute-setting time! Let's try to parse the value. */
+    const parsed = parseJson(value) ?? parseDeg(value)
+
+    /* Handle properties that provide .set methods */
+    if (object[key]?.set !== undefined) {
+      /* If the value is an array, feed its destructured representation to the set method. */
+      if (Array.isArray(parsed)) {
+        object[key].set(...parsed)
+        return
+      }
+
+      /* A bit of special handling for "scale" properties, where we'll also accept a single numerical value */
+      if (key === "scale" && typeof parsed === "number") {
+        object[key].setScalar(parsed)
+        return
+      }
+
+      /* If we have a parsed value, set it directly */
+      if (parsed) {
+        object[key].set(parsed)
+        return
+      }
+
+      /* Otherwise, set the original string value, but split by commas */
+      const list = attributeValueToArray(value)
+      object[key].set(...list)
+      return
+    }
+
+    /*
+    Is the property an object? In that case, we'll assume that the string value of the attribute
+    contains a DOM selector that references another object that we should assign here.
+    */
+    if (typeof object[key] === "object") {
+      const referencedObject = getThreeObjectBySelector(value)
+
+      if (referencedObject) {
+        object[key] = referencedObject
+      } else {
+        console.error(`No element was found for selector "${value}". 😭`)
+      }
+
+      return
+    }
+
+    /*
+    If we've reached this point, we're finally able to set a property on the object.
+    Amazing! But let's only do it if the property key is actually known.
+    */
+    if (key in object) object[key] = parsed !== undefined ? parsed : value
   }
+}
+
+const parseJson = (value: string) => {
+  let parsed = undefined
+
+  try {
+    parsed = JSON.parse(value)
+  } catch (e) {}
+
+  return parsed
 }